# [3.x] convert AST_Seq from binary tree to array