… lib/compress.js