@alexlamsl 2017-02-02
### master

```
https://code.jquery.com/jquery-3.1.1.js
Timing information (compressed 1 files):
- parse: 0.105s
- scope: 0.297s
- squeeze: 2.019s
- mangle: 0.031s
- generate: 0.101s


https://code.angularjs.org/1.6.1/angular.js
Timing information (compressed 1 files):
- parse: 0.236s
- scope: 0.556s
- squeeze: 2.889s
- mangle: 0.065s
- generate: 0.201s


https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
Timing information (compressed 1 files):
- parse: 0.463s
- scope: 1.496s
- squeeze: 53.221s
- mangle: 0.144s
- generate: 0.550s


https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js
Timing information (compressed 1 files):
- parse: 0.041s
- scope: 0.103s
- squeeze: 0.480s
- mangle: 0.013s
- generate: 0.036s
```

### This PR

```
https://code.jquery.com/jquery-3.1.1.js
Timing information (compressed 1 files):
- parse: 0.106s
- scope: 0.289s
- squeeze: 1.777s
- mangle: 0.042s
- generate: 0.114s


https://code.angularjs.org/1.6.1/angular.js
Timing information (compressed 1 files):
- parse: 0.223s
- scope: 0.576s
- squeeze: 2.859s
- mangle: 0.066s
- generate: 0.190s


https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
Timing information (compressed 1 files):
- parse: 0.461s
- scope: 1.529s
- squeeze: 53.390s
- mangle: 0.145s
- generate: 0.524s


https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js
Timing information (compressed 1 files):
- parse: 0.039s
- scope: 0.118s
- squeeze: 0.485s
- mangle: 0.011s
- generate: 0.037s
```