@kzc 2017-02-02
Wow! Nicely done.

I trust that it works. I'm not going to pretend to review it.

I'd suggest to change the AST node name to something other than AST_Seq since it's a departure from the old AST. AST_Sequence, perhaps?

You'd likely have to make an equivalent patch for the `harmony` branch since it's unlikely to merge cleanly.  Would be best to wait for the PR queue to emptied and for a release made before rebasing.