@kzc 2017-02-02
@alexlamsl If I'm interpreting your numbers correctly, there is no speed improvement?