@alexlamsl 2017-02-02
@kzc jQuery seems to benefit from this change, whilst other libraries do not suffer noticeably from any performance regressions.

I need to dig out https://github.com/mishoo/UglifyJS2/issues/1145#issuecomment-227006941 to test against this PR next, but let me take care of `run-test.js` as per your request first 😜