@kzc 2017-02-02
Didn't look at the PR in too much detail, but are there any asserts preventing a sequence array length of 1?