@alexlamsl 2017-02-02
Not at the moment, no. I'm thinking about how to consolidate all the constructions in order to perform the check reliably.