@alexlamsl 2017-02-02
Not only just `length > 1`, but also none of the elements being `AST_Seq` as well.