@alexlamsl 2017-02-02
Replaced all occurrences of `[].forEach()` in `ast.js` so far...

```
https://code.jquery.com/jquery-3.1.1.js
Timing information (compressed 1 files):
- parse: 0.105s
- scope: 0.292s
- squeeze: 1.798s
- mangle: 0.031s
- generate: 0.105s


https://code.angularjs.org/1.6.1/angular.js
Timing information (compressed 1 files):
- parse: 0.230s
- scope: 0.558s
- squeeze: 2.742s
- mangle: 0.063s
- generate: 0.183s


https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
Timing information (compressed 1 files):
- parse: 0.492s
- scope: 1.419s
- squeeze: 49.142s
- mangle: 0.136s
- generate: 0.523s


https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js
Timing information (compressed 1 files):
- parse: 0.040s
- scope: 0.099s
- squeeze: 0.475s
- mangle: 0.011s
- generate: 0.037s
```

So even `angular.js` is starting to get noticeably faster. By the way, this is `--stats` from `bin\uglifyjs` as I was too lazy to time stuff myself 😝