@kzc 2017-02-02
Does the acronym `[BC]` mean "Binary Compatibility" change or "Breaking Change" or other?