@alexlamsl 2017-02-02
I'd be grateful if you can run `node test\benchmark.js` on your system to validate these timings. I'm on `node-chakracore` and presumably you are on the standard V8.