@kzc 2017-02-02
> Replaced all occurrences of [].forEach() in ast.js so far...
> So even angular.js is starting to get noticeably faster

From your stats, compress appears to be roughly 5% faster with that trivial `forEach` -> `for` change - good stuff.

I should have asked you to change it sooner!  ;-)