@alexlamsl 2017-02-02
Let's check V8 and/or SpiderMonkey before confirming this 👻 

(Re-opening against Travis CI time-out.)