@kzc 2017-02-02
test/benchmark.js timings using node v6.9.0 on Mac:

**master**:
```
https://code.jquery.com/jquery-3.1.1.js
Timing information (compressed 1 files):
- parse: 0.152s
- scope: 0.300s
- squeeze: 1.777s
- mangle: 0.026s
- generate: 0.091s

https://code.angularjs.org/1.6.1/angular.js
Timing information (compressed 1 files):
- parse: 0.334s
- scope: 0.471s
- squeeze: 2.414s
- mangle: 0.050s
- generate: 0.152s

https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
Timing information (compressed 1 files):
- parse: 0.580s
- scope: 1.100s
- squeeze: 31.003s
- mangle: 0.080s
- generate: 0.450s

https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js
Timing information (compressed 1 files):
- parse: 0.079s
- scope: 0.107s
- squeeze: 0.477s
- mangle: 0.017s
- generate: 0.043s
```
**this PR with `forEach` to `for` change 02c5ca6b58cde2d23c965c71f7eb251695169bd0:**
```
https://code.jquery.com/jquery-3.1.1.js
Timing information (compressed 1 files):
- parse: 0.164s
- scope: 0.265s
- squeeze: 1.486s
- mangle: 0.024s
- generate: 0.089s

https://code.angularjs.org/1.6.1/angular.js
Timing information (compressed 1 files):
- parse: 0.312s
- scope: 0.479s
- squeeze: 2.326s
- mangle: 0.048s
- generate: 0.160s

https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
Timing information (compressed 1 files):
- parse: 0.579s
- scope: 1.160s
- squeeze: 27.843s
- mangle: 0.075s
- generate: 0.370s

https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js
Timing information (compressed 1 files):
- parse: 0.073s
- scope: 0.103s
- squeeze: 0.472s
- mangle: 0.008s
- generate: 0.045s
```
math.js compress is 10% quicker.