@alexlamsl 2017-02-02
Cool! Let me finish off `compress.js` and `output.js` and test again 😉