@kzc 2017-02-02
uglify-js@2.7.5:
```
$ time uglifyjs math.js -m -c warnings=false | shasum
2b42f6939f4bc5f20350d74e5c385a535d1876fa  -

real	0m12.760s
user	0m13.298s
sys	0m0.164s
```
This PR (02c5ca6b58cde2d23c965c71f7eb251695169bd0):
```
$ time bin/uglifyjs math.js -m -c warnings=false | shasum
2b42f6939f4bc5f20350d74e5c385a535d1876fa  -

real	0m11.868s
user	0m12.273s
sys	0m0.163s
```
Same output shasums.
7% faster.