@kzc 2017-02-02
@alexlamsl Can you try timing that pathological massive sequence test case against master and this PR?