@alexlamsl 2017-02-02
Will do once I sort out this curious case of https://github.com/mishoo/UglifyJS2/blob/0610c020b1544820be9898a285ab6c9066490552/lib/compress.js#L2080

Replacing it with a for-loop:
```js
for (var i = 0, len = this.definitions; i < len; i++) {
    this.definitions[i].value = null;
}
```
... gives me test failures. I must be staring at something obvious,,,.