@kzc 2017-02-02
`len = this.definitions` should be `len = this.definitions.length`